requirements.txt
# pip install -r requirements.txt
# or
# pip freeze > requirements.txt

# pip list

Flask==3.0.0
google-cloud-dlp==3.13.0
cryptography==41.0.5
PyMuPDF==1.23.6
